Updtaed read me setup file

+++ circleci-scripts/readme-setup.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
 pwd
-rm README.md
+if test -f "./README.md"; then
+    echo "file exits"
+    rm README.md
+fi
+
 cd readme-sources
 ls
 mdmerge main.md coverage.md -o README.md
